+++ project/CHANGELOG.md
@@ -5,3 +5,11 @@ All notable changes to this project will be documented in this file.
 ### Added
 * Project creation
 * Project configuration
+
+## [v0.0.2] - 2020-03-01
+### Added
+* Chocolatey on Windows
+
+## [v0.1.0] - 2020-04-24
+### Changed
+* Integrated chocolatey architecture